layer3 refactoring done
