+++ cla/shared/src/main/scala/com/colingodsey/logos/cla/L3Region.scala
@@ -9,47 +9,7 @@ trait Layer extends DutyCycle.Booster {
   implicit val config: CLA.Config
 }
 
-trait SequenceLayer extends Layer {
-  import CLA._
 
-  type ColumnType <: Column
-
-  def columns: IndexedSeq[ColumnType]
-  def getLearningCells: Stream[Cell]
-  //def input(idx: Int): Boolean
-  def inhibitionRadius: Double
-
-  def columnsNear(loc: Location, rad: Radius) = {
-    val min = math.max(0, loc - rad).toInt
-    val max = math.min(config.regionWidth - 1, loc + rad).toInt
-
-    (min to max).iterator map columns
-  }
-
-  def numActiveFromPrediction = columns.count(column => column.active && column.wasPredicted)
-  def numActive = columns.count(_.active)
-
-  def anomalyScore = {
-    val na = numActive.toDouble
-
-    if(na == 0) 0.0
-    else 1.0 - numActiveFromPrediction / na
-  }
-
-  def averageReceptiveFieldRadius = {
-    //TODO: filter or not?
-    val filtered = columns//.filter(_.active)
-
-    val sum = filtered.map(_.receptiveFieldRadius).sum
-
-    sum.toDouble / filtered.length
-  }
-
-  def getLearningCells(c: Column): Stream[Cell]  =
-    getLearningCells.filter(_.column != c)
-
-  def activeColumns = columns.iterator.filter(_.active)
-}
 
 trait L3Layer extends SequenceLayer {
   override type ColumnType = L3Column
@@ -91,6 +51,7 @@ class L3Region(implicit val config: CLA.Config,
   }
 
   var inhibitionRadiusAverage = RollingAverage(dutyAverageFrames)
+
   def maxDutyCycle = inputLayer.maxDutyCycle
 
   inhibitionRadiusAverage += regionWidth / 2.0
